% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_SPD.R
\name{methods_SPD}
\alias{methods_SPD}
\title{Returns the matrix of eigenvectors computed from the stepwise estimation of eigenvectors, joint diagonalization based on the FG algorithm, or from the eigendecomposition of the sum of SPD matrices.}
\usage{
methods_SPD(grads, n, m, d, method = c("SSPD", "SEE", "FG"))
}
\arguments{
\item{grads}{Gradient matrix, where the columns are the concatenation of the objective-wise gradients}

\item{n}{Sample size}

\item{m}{Number of objectives}

\item{d}{Number of input dimensions}

\item{method}{Name of the SPD method to choose from
\itemize{
\item 'SSPD' to compute the eigenvectors from the sum of SPD matrices,
\item 'SEE' to compute the eigenvectors in a stepwise fashion, or
\item 'FG' to do simultaneous diagonalization using the Flury and Gautschi algorithm.
}}
}
\value{
Matrix of eigenvectors (column vectors)
}
\description{
Returns the matrix of eigenvectors computed from the stepwise estimation of eigenvectors, joint diagonalization based on the FG algorithm, or from the eigendecomposition of the sum of SPD matrices.
}
\examples{
fn <- problem_car_side_impact
d <- 7
n <- 1000
X <- matrix(runif(d * n), n)
Y <- t(apply(X, 1, fn))
grads <- t(apply(X, 1, function(x) c(t(numDeriv::jacobian(fn, x = x)))))
methods_SPD(grads, n, ncol(Y), d, method = "FG")

}
\references{
\itemize{
\item{Flury, B. N. and W. Gautschi (1986). An algorithm for simultaneous orthogonal transformation of several positive definite symmetric matrices to nearly diagonal form. SIAM Journal on Scientific and Statistical Computing 7 (1), 169–184}
\item{Trendafilov, N. T. (2010). Stepwise estimation of common principal components. Computational Statistics & Data Analysis 54 (12), 3446–3457.}
}
}
