% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_convex_hull.R
\name{min_convex_hull}
\alias{min_convex_hull}
\title{A recursive algorithm to compute the minimum L2-norm point in a polytope.}
\usage{
min_convex_hull(P)
}
\arguments{
\item{P}{Matrix for the convex hull based on which to compute the minimum L2-norm element}
}
\value{
Vector representing the minimum L2-norm point of the convex hull of \code{P}
}
\description{
A recursive algorithm to compute the minimum L2-norm point in a polytope.
}
\examples{
X <- matrix(runif(3*10), 10)
min_convex_hull(X)

}
\references{
Sekitani, K. and Y. Yamamoto (1993). A recursive algorithm for finding the minimum
norm point in a polytope and a pair of closest points in two polytopes. Mathematical Programming 61, 233–249
}
