% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_car.R
\name{problem_car_side_impact}
\alias{problem_car_side_impact}
\title{Car side-impact problem}
\usage{
problem_car_side_impact(
  x,
  params = list(p = NULL, mp = c(0.345, 0.192, 0, 0), sp = c(0.06, 0.06, 10, 10), fixed =
    TRUE)
)
}
\arguments{
\item{x}{a vector in [0, 1]^7 corresponding to
1: Thickness of B-Pillar inner
2: Thickness of B-Pillar reinforcement
3: Thickness of floor side inner
4: Thickness of cross members
5: Thickness of door beam
6 : Thickness of door beltline reinforcement
7 : Thickness of roof rail}

\item{params}{List of additional parameters:
\itemize{
\item {
\code{p} optional vector of values for the random parameter, if \code{fixed = TRUE}, no sampling is done and they are fixed at their mean:
8: Material of B-Pillar inner
9: Material of floor side inner
10: Barrier height
11: Barrier hitting position}
\item{\code{mp,sp} mean and standard deviation for p}
\item{\code{fixed} Logical}
}}
}
\value{
Vector of size 11
}
\description{
Car side-impact problem
}
\examples{
d <- 7
n <- 1000
X <- matrix(runif(d*n), n)
Y <- t(apply(X, 1, problem_car_side_impact))

pairs(Y)

}
\references{
Deb, K. et al. “Reliability-Based Optimization Using Evolutionary Algorithms.”
Evolutionary Computation, IEEE Transactions on 13.5 (2009): 1054-1074. 2009 Institute of
Electrical and Electronics Engineers
}
