% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_marinedes.R
\name{problem_marinedes}
\alias{problem_marinedes}
\title{Conceptual marine design problem}
\usage{
problem_marinedes(x, params = list(constraints = FALSE))
}
\arguments{
\item{x}{Vector in [0,1]^6}

\item{params}{List of additional parameters:
\itemize{
\item{\code{returnCST}, logical, indicating whether the constraints are returned in the output}
}}
}
\value{
If returnCST is \code{TRUE} a vector of size 12, the first 3 are the objectives, next are the constraints.
Otherwise, it returns three objectives.
}
\description{
Conceptual marine design problem
}
\note{
Adapted from Tanabe, R., & Ishibuchi, H. (2020). An easy-to-use real-world multi-objective optimization problem suite. Applied Soft Computing, 89, 106078.
}
\examples{
d <- 6
n <- 1000
X <- matrix(runif(d*n), n)
Y <- t(apply(X, 1, problem_marinedes))

pairs(Y[, 1:3])
}
\references{
\itemize{
\item{M. G. Parsons and R. L. Scott, "Formulation of Multicriterion Design Optimization Problems for Solution With Scalar Numerical Optimization Methods," J. Ship Research, vol. 48, no. 1, pp. 61-76, 2004.}
\item{Tanabe, R., & Ishibuchi, H. (2020). An easy-to-use real-world multi-objective optimization problem suite. Applied Soft Computing, 89, 106078.}
}
}
