% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_penicillin.R
\name{problem_penicillin}
\alias{problem_penicillin}
\title{Penicillin production test function based on the implementation github.com/HarryQL/TuRBO-Penicillin.}
\usage{
problem_penicillin(input, params = list(t = 100, returnCST = FALSE))
}
\arguments{
\item{input}{Matrix of input data}

\item{params}{List of additional parameters:
\itemize{
\item \code{t} Reaction time in hours, defaulted to 100
\item \code{returnCST} Optional. If \code{TRUE} returns the constraints as part of the output.
}}
}
\value{
Vector of objective values: if \code{returnCST} is set to \code{TRUE} a vector of size 5 where the last 3 columns correspond to the constraints.
}
\description{
Penicillin production test function based on the implementation github.com/HarryQL/TuRBO-Penicillin.
}
\references{
Liang, Q. and L. Lai (2021). Scalable bayesian optimization accelerates process optimization of penicillin production. In NeurIPS 2021 AI for Science Workshop.
}
