% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_switch_ripple.R
\name{problem_switch_ripple}
\alias{problem_switch_ripple}
\title{Switching ripple test function.
Corresponds to the switching ripple suppressor design problem for voltage source inversion in powered system.}
\usage{
problem_switch_ripple(x, params = list(returnCST = FALSE))
}
\arguments{
\item{x}{vector specifying the location where the function is to be evaluated, of size k + 4, k > 1, see Details.}

\item{params}{List of additional parameters:
\itemize{
\item \code{returnCST} Optional. If set to \code{TRUE} the last 5 columns in the output correspond to the values of the constraints.
}}
}
\value{
Vector of objectives. The first \code{k} values are related to the suppression of harmonics while the \code{k+1} one is the sum of inductors.
}
\description{
Switching ripple test function.
Corresponds to the switching ripple suppressor design problem for voltage source inversion in powered system.
}
\details{
Columns of \code{x} correspond to L1, L2, L3, C1, ..., Ck, Cf where k is an arbitrary integer > 1.

Parameters of the problems follow Table 2 in (Zhang et al, 2019).
}
\references{
\itemize{
\item{ Zhang, Z., He, C., Ye, J., Xu, J., & Pan, L. (2019). Switching ripple suppressor design of the grid-connected inverters: A perspective of many-objective optimization with constraints handling. Swarm and evolutionary computation, 44, 293-303.}
\item{ He, C., Tian, Y., Wang, H., & Jin, Y. (2019). A repository of real-world datasets for data-driven evolutionary multiobjective optimization. Complex & Intelligent Systems, 1-9.}
}
}
