% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_synthetic.R
\name{problem_synthetic}
\alias{problem_synthetic}
\title{Two-dimensional tri-objective problem based on Branin and Currin functions.}
\usage{
problem_synthetic(x, params = list(A = NULL, B = NULL, seed = 19))
}
\arguments{
\item{x}{Vector or matrix to evaluate at, takes values in [0,1]^3}

\item{params}{List of additional parameters:
\itemize{
\item \code{A} and \code{B} embedding matrices of size 3x3
\item \code{seed} random seed for reproducibility
}}
}
\description{
Two-dimensional tri-objective problem based on Branin and Currin functions.
}
\note{
Problem generated by sampling 2 random matrices and extracting an orthonormal basis from them.
}
